break up big molecular data query form file
